修正jpg 主键类型

+++ micro-tf-framework/src/main/java/com/microtf/framework/jpa/ArticleRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author guliuzhong
  */
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
 }
